Add files via upload
This case study aims to give you an idea of applying EDA in a real business scenario. In this case study, we will apply some of the concepts of risk analytics in banking and financial services and understand how to minimize the risk of losing money while lending to customers.